requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-google==10.17.0